refactor(templates): extract shared widgets to templates.core.widgets
Move the family-neutral widgets — CardWidget, TableWidget, TimelineAxisWidget —
and the Divider / AccentStrip / Spacer decoration primitives into
com.demcha.compose.document.templates.core.widgets so any template family can
reuse them, and retire the now-empty templates.widgets and templates.decorations
packages. Repoint the cv.v2 / coverletter.v2 callers and relocate the widget and
decoration test mirrors. Pure move — render byte-identical (1612 tests, zero
snapshot updates); a deliberate binary break on the 2.0 line (japicmp report-only).
