+++ src/main/java/com/demcha/compose/document/templates/core/widgets/package-info.java
@@ -1,0 +1,21 @@
+/**
+ * <h2>Core — shared visual widgets (widgets)</h2>
+ *
+ * <p>Family-neutral, composable visual building blocks reusable by every
+ * template family — they make no CV / invoice / proposal assumptions and read
+ * only the document DSL plus a {@code BrandTheme}'s tokens.</p>
+ *
+ * <ul>
+ *   <li>{@link com.demcha.compose.document.templates.core.widgets.CardWidget}
+ *       — a rounded container shell.</li>
+ *   <li>{@link com.demcha.compose.document.templates.core.widgets.TableWidget}
+ *       — a fixed / grid table.</li>
+ *   <li>{@link com.demcha.compose.document.templates.core.widgets.TimelineAxisWidget}
+ *       — a line-and-marker axis.</li>
+ *   <li>{@link com.demcha.compose.document.templates.core.widgets.Divider} /
+ *       {@link com.demcha.compose.document.templates.core.widgets.AccentStrip} /
+ *       {@link com.demcha.compose.document.templates.core.widgets.Spacer}
+ *       — decoration primitives a preset can attach between blocks.</li>
+ * </ul>
+ */
+package com.demcha.compose.document.templates.core.widgets;
