+++ src/main/java/com/demcha/compose/document/templates/coverletter/v2/presets/PanelLetter.java
@@ -15,7 +15,7 @@ import com.demcha.compose.document.templates.core.text.TextStyles;
 import com.demcha.compose.document.templates.core.text.TextOrnaments;
 import com.demcha.compose.document.templates.cv.v2.data.CvIdentity;
 import com.demcha.compose.document.templates.core.theme.BrandTheme;
-import com.demcha.compose.document.templates.widgets.CardWidget;
+import com.demcha.compose.document.templates.core.widgets.CardWidget;
 import com.demcha.compose.font.FontName;
 
 import java.util.Locale;
