--- src/main/java/com/demcha/compose/document/templates/decorations/package-info.java
@@ -1,29 +1,0 @@
-/**
- * Templates v2 decoration library — dividers, accent strips, panels, banners, ornaments.
- *
- * <p>Decorations are the visual artefacts (lines, strips, soft panels,
- * coloured banners) that templates can attach between modules, around a
- * header, or as the start/end of a module body. They are first-class
- * components rather than baked-in composer logic, so any preset can mix
- * and match them and consumers can reuse them in their own templates.</p>
- *
- * <p>This package will be populated during Phase B of the Templates v2
- * migration with at least:</p>
- *
- * <ul>
- *   <li>{@code Divider} — thin / thick / dashed / dotted-accent /
- *       gradient-fade / ornamental horizontal lines.</li>
- *   <li>{@code AccentStrip} — left / top / bottom / corner coloured
- *       strips for emphasising a region.</li>
- *   <li>{@code Panel} — soft / bordered / glass background panels.</li>
- *   <li>{@code Banner} — full-width / angled / gradient banners
- *       (typical for the BlueBanner-style CV pattern).</li>
- *   <li>{@code Ornament} — small marks (bracket, quote, icon) for
- *       editorial flourishes.</li>
- *   <li>{@code Spacer} — semantic vertical spacing (small / medium /
- *       large / custom).</li>
- * </ul>
- *
- * @since 1.6.0
- */
-package com.demcha.compose.document.templates.decorations;
