--- src/main/java/com/demcha/compose/document/templates/widgets/package-info.java
@@ -1,14 +1,0 @@
-/**
- * Shared visual widgets available to all template families.
- *
- * <p>These widgets sit above the raw document DSL and below
- * preset-specific composition. They are deliberately generic: CV,
- * proposal, invoice, cover-letter, and future templates can reuse
- * them without depending on a CV-only package.</p>
- *
- * <p>Current shared widgets include configurable tables
- * ({@link com.demcha.compose.document.templates.widgets.TableWidget})
- * and reusable card/container shells
- * ({@link com.demcha.compose.document.templates.widgets.CardWidget}).</p>
- */
-package com.demcha.compose.document.templates.widgets;
